requirements.txt
# CBMCD Requirements
# Cultural Bias Mitigation with Context-Dependent preference

# Core
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0

# LoRA
peft>=0.7.0

# Quantization (optional)
bitsandbytes>=0.41.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Progress bars
tqdm>=4.66.0

# Logging (optional)
wandb>=0.15.0

# Development
black>=23.0.0
isort>=5.12.0
pytest>=7.4.0
